Add changes to meet PR requirements
