+++ index.html
@@ -13,13 +13,13 @@
   <section id="booksDisplay"></section>
   <form action="#" id="bookForm">
     <br>
-    <input type="text" placeholder="Book title" required /> <br />
-    <br />
-    <input type="text" placeholder="Author" required /> <br />
-    <br />
+    <input type="text" placeholder="Book title" required><br>
+    <br>
+    <input type="text" placeholder="Author" required> <br>
+    <br>
     <button type="submit" id="submitBttn">Add</button>
   </form>
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 
 </html>
