+++ js/modules/Books.js
@@ -1,0 +1,13 @@
+export default class Books {
+  static ids = 0;
+
+  static booksData = [];
+
+  static initializeBooks(data) {
+    this.booksData = data;
+  }
+
+  static getBooks() {
+    return this.booksData;
+  }
+}
